renamed react component to Notificami
