+++ packages/notificami-react-components/src/index.js
@@ -1,5 +1,5 @@
-export { NotificationsProvider } from './components/NotificationsProvider'
-export { NotificationsWidget } from './components/NotificationsWidget'
+export { NotificamiProvider } from './components/NotificamiProvider'
+export { NotificamiWidget } from './components/NotificamiWidget'
 export { NotificationsBox } from './components/NotificationsBox'
 export { NotificationsList } from './components/NotificationsList'
 export * from './nesClient'
